maturity allowlist as separate
