+++ src/pages/Admin/Charity/Maturity/MaturityDate/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./MaturityDate";
